+++ CHANGES.md
@@ -1,3 +1,7 @@
 # 0.0.1
 
-Bootstrapped the project.
+Bootstrapped the project:
+
+* Add common extensible event type
+* Add handle registration with `Telemetry.attach`
+* Add event emitting with `Telemetry.emit`